+++ packages/article-image/__tests__/article-image.test.js
@@ -1,6 +1,0 @@
-/* eslint-env jest */
-
-import "react-native";
-import shared from "./shared";
-
-shared();