+++ packages/article/article.web.js
@@ -1,14 +1,16 @@
 import React from "react";
+import { View } from "react-native";
 import PropTypes from "prop-types";
-import get from "lodash.get";
-import { renderTrees } from "@times-components/markup";
-import ArticleImage from "@times-components/article-image";
 import { AdComposer } from "@times-components/ad";
-import { withTrackingContext } from "@times-components/tracking";
-import PullQuote from "@times-components/pull-quote";
 
 import ArticleError from "./article-error";
 import ArticleLoading from "./article-loading";
+import ArticleHeader from "./article-header";
+import ArticleMeta from "./article-meta";
+import ArticleBody from "./article-body";
+import LeadAssetComponent from "./article-lead-asset.web";
+
+import articleTrackingContext from "./article-tracking-context";
 
 import {
   MainContainer,
@@ -16,23 +18,6 @@ import {
   MetaContainer,
   BodyContainer
 } from "./styles/responsive";
-
-import {
-  ParagraphContainer,
-  Paragraph,
-  LeadAsset,
-  LeadAssetMobile,
-  LeadAssetDesktop,
-  MediaContainerMobile,
-  MediaContainerDesktop,
-  PrimaryImg,
-  SecondaryImg,
-  InlineImg,
-  PullQuoteContainer,
-  PullQuoteResp
-} from "./styles/body/responsive";
-import ArticleHeader from "./article-header";
-import ArticleMeta from "./article-meta";
 
 const withAdComposer = (children, section = "article") => (
   <AdComposer section={section}>{children}</AdComposer>
@@ -48,19 +33,13 @@ class ArticlePage extends React.Component {
       byline,
       publishedTime,
       publicationName,
-      leadAsset
+      leadAsset,
+      content
     } = articleData;
-    const LeadAssetMedia = ArticlePage.renderLeadAsset(leadAsset);
-    const contentArray = articleData.content.map((data, index) => ({
-      data,
-      index
-    }));
-    const BodyView = contentArray.map((content, index) =>
-      ArticlePage.renderBody(content, index)
-    );
+
     return (
       <MainContainer>
-        <MediaContainerMobile>{LeadAssetMedia}</MediaContainerMobile>
+        <LeadAssetComponent device="MOBILE" leadAsset={leadAsset} />
         <HeaderContainer>
           <ArticleHeader
             headline={headline}
@@ -69,7 +48,7 @@ class ArticlePage extends React.Component {
             label={label}
           />
         </HeaderContainer>
-        <BodyContainer>
+        <View>
           <MetaContainer>
             <ArticleMeta
               byline={byline}
@@ -77,100 +56,13 @@ class ArticlePage extends React.Component {
               publicationName={publicationName}
             />
           </MetaContainer>
-          <MediaContainerDesktop>{LeadAssetMedia}</MediaContainerDesktop>
-          {BodyView}
-        </BodyContainer>
+          <LeadAssetComponent device="DESKTOP" leadAsset={leadAsset} />
+          <BodyContainer>
+            <ArticleBody content={content} />
+          </BodyContainer>
+        </View>
       </MainContainer>
     );
-  }
-
-  static renderLeadAsset(leadAsset) {
-    return (
-      <LeadAsset>
-        <LeadAssetMobile>
-          <ArticleImage
-            imageOptions={{
-              display: "",
-              ratio: leadAsset.crop.ratio,
-              url: leadAsset.crop.url
-            }}
-          />
-        </LeadAssetMobile>
-        <LeadAssetDesktop>
-          <ArticleImage
-            imageOptions={{
-              display: "",
-              ratio: leadAsset.crop.ratio,
-              url: leadAsset.crop.url
-            }}
-            captionOptions={{
-              caption: leadAsset.caption,
-              credits: leadAsset.credits
-            }}
-          />
-        </LeadAssetDesktop>
-      </LeadAsset>
-    );
-  }
-
-  static renderBody(content) {
-    return renderTrees([content.data], {
-      paragraph(key, attributes, children) {
-        return (
-          <ParagraphContainer
-            testID={`paragraph-${content.index}`}
-            accessibilityLabel={`paragraph-${content.index}`}
-            key={key}
-          >
-            <Paragraph>{children}</Paragraph>
-          </ParagraphContainer>
-        );
-      },
-      image(key, attributes) {
-        const ImageContainer = ArticlePage.getImageContainer(
-          attributes.display
-        );
-        return (
-          <ImageContainer key={key}>
-            <ArticleImage
-              imageOptions={{
-                display: attributes.display,
-                ratio: attributes.ratio,
-                url: attributes.url
-              }}
-              captionOptions={{
-                caption: attributes.caption,
-                credits: attributes.credits
-              }}
-            />
-          </ImageContainer>
-        );
-      },
-      pullquote(key, attributes) {
-        return (
-          <PullQuoteContainer>
-            <PullQuoteResp>
-              <PullQuote
-                key={key}
-                content={attributes.content}
-                caption={attributes.caption.name}
-              />
-            </PullQuoteResp>
-          </PullQuoteContainer>
-        );
-      }
-    });
-  }
-
-  static getImageContainer(imageType) {
-    switch (imageType) {
-      case "secondary":
-        return SecondaryImg;
-      case "inline":
-        return InlineImg;
-      default:
-        return PrimaryImg;
-    }
   }
 
   render() {
@@ -210,11 +102,4 @@ ArticlePage.defaultProps = {
   error: null
 };
 
-export default withTrackingContext(ArticlePage, {
-  trackingObject: "Article",
-  getAttrs: ({ article } = {}) => ({
-    byline: get(article, "byline[0].children[0].attributes.value", ""),
-    headline: get(article, "headline", ""),
-    publishedTime: get(article, "publishedTime", "")
-  })
-});
+export default articleTrackingContext(ArticlePage);