+++ packages/article/styles/responsive.js
@@ -37,6 +37,10 @@ export const MetaContainer = styled(View)`
 `;
 
 /* --- Body --- */
+
+const leadAssetBehaviour = leadAsset => (leadAsset ? "0" : "-9px");
+
 export const BodyContainer = styled(View)`
   display: block;
+  margin-top: ${props => leadAssetBehaviour(props.leadAsset)};
 `;