+++ packages/author-head/__tests__/ios/jest.config.js
@@ -1,3 +1,9 @@
 const jestConfigurator = require("@times-components/jest-configurator");
 
-module.exports = jestConfigurator("author-head", "ios");
+module.exports = Object.assign(jestConfigurator("author-head", "ios"), {
+  collectCoverageFrom: [
+    "**/packages/author-head/*.js",
+    "!**/packages/author-head/*.web.js",
+    "!**/packages/author-head/*.stories.js"
+  ]
+});