+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.3"></a>
+## [0.12.3](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.2...@times-components/pagination@0.12.3) (2017-12-13)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
+<a name="0.12.2"></a>
+## [0.12.2](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.0...@times-components/pagination@0.12.2) (2017-12-12)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
+<a name="0.12.1"></a>
+## [0.12.1](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.0...@times-components/pagination@0.12.1) (2017-12-12)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
 <a name="0.12.0"></a>
 # [0.12.0](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.11.5...@times-components/pagination@0.12.0) (2017-12-12)